+++ src/config.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import * as utils from "./utils";
-import { ConfigMembers } from "./types";
+import { ConfigMembers, RefreshType } from "./types";
 import { Constants } from "./constants";
 import { LeoIntegration } from "./leoIntegration";
 
@@ -12,12 +12,14 @@ export class Config implements ConfigMembers {
     // TODO : Should use default values in a centralized way
     public checkForChangeExternalFiles: string = "none";  // Used in leoBridge script
     public defaultReloadIgnore: string = "none"; // Used in leoBridge script
+    public leoTreeBrowse: boolean = false;
     public treeKeepFocus: boolean = true;
     public treeKeepFocusWhenAside: boolean = false;
     public treeInExplorer: boolean = true;
     public showOpenAside: boolean = true;
     public statusBarString: string = "";
     public statusBarColor: string = "";
+    public showEditOnNodes: boolean = true;
     public showArrowsOnNodes: boolean = false;
     public showAddOnNodes: boolean = false;
     public showMarkOnNodes: boolean = false;
@@ -31,6 +33,7 @@ export class Config implements ConfigMembers {
     public connectionPort: number = Constants.TCPIP_DEFAULT_PORT;
 
     private _isSettingConfig: boolean = false;
+    private _needsTreeRefresh: boolean = false;
 
     constructor(
         private _context: vscode.ExtensionContext,
@@ -45,12 +48,15 @@ export class Config implements ConfigMembers {
         return {
             checkForChangeExternalFiles: this.checkForChangeExternalFiles,  // Used in leoBridge script
             defaultReloadIgnore: this.defaultReloadIgnore,  // Used in leoBridge script
+
+            leoTreeBrowse: this.leoTreeBrowse,
             treeKeepFocus: this.treeKeepFocus,
             treeKeepFocusWhenAside: this.treeKeepFocusWhenAside,
             treeInExplorer: this.treeInExplorer,
             showOpenAside: this.showOpenAside,
             statusBarString: this.statusBarString,
             statusBarColor: this.statusBarColor,
+            showEditOnNodes: this.showEditOnNodes,
             showArrowsOnNodes: this.showArrowsOnNodes,
             showAddOnNodes: this.showAddOnNodes,
             showMarkOnNodes: this.showMarkOnNodes,
@@ -74,8 +80,11 @@ export class Config implements ConfigMembers {
         this._isSettingConfig = true;
         const w_promises: Thenable<void>[] = [];
         const w_vscodeConfig = vscode.workspace.getConfiguration(Constants.CONFIG_NAME);
-
         p_changes.forEach(i_change => {
+            if (i_change && i_change.code.includes(Constants.CONFIG_REFRESH_MATCH)) {
+                // Check if tree refresh is required for hover-icons to be displayed or hidden accordingly
+                this._needsTreeRefresh = true;
+            }
             if (w_vscodeConfig.inspect(i_change.code)!.defaultValue === i_change.value) {
                 // set as undefined - same as default
                 w_promises.push(w_vscodeConfig.update(i_change.code, undefined, true));
@@ -87,6 +96,12 @@ export class Config implements ConfigMembers {
             }
         });
         return Promise.all(w_promises).then(() => {
+            if (this._needsTreeRefresh) {
+                this._needsTreeRefresh = false;
+                setTimeout(() => {
+                    this._leoIntegration.configTreeRefresh();
+                }, 200);
+            }
             this._isSettingConfig = false;
             this.buildFromSavedSettings();
         });
@@ -114,14 +129,17 @@ export class Config implements ConfigMembers {
             // TODO : Should use default values in a centralized way
             this.checkForChangeExternalFiles = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.CHECK_FOR_CHANGE_EXTERNAL_FILES, "none");
             this.defaultReloadIgnore = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.DEFAULT_RELOAD_IGNORE, "none");
+            this.leoTreeBrowse = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.LEO_TREE_BROWSE, true);
             this.treeInExplorer = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.TREE_IN_EXPLORER, true);
             this.showOpenAside = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.SHOW_OPEN_ASIDE, true);
+            this.showEditOnNodes = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.SHOW_EDIT, false);
             this.showArrowsOnNodes = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.SHOW_ARROWS, false);
             this.showAddOnNodes = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.SHOW_ADD, false);
             this.showMarkOnNodes = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.SHOW_MARK, false);
             this.showCloneOnNodes = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.SHOW_CLONE, false);
             this.showCopyOnNodes = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.SHOW_COPY, false);
             // * Interface settings
+            this.leoTreeBrowse = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.LEO_TREE_BROWSE, true);
             this.treeKeepFocus = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.TREE_KEEP_FOCUS, true);
             this.treeKeepFocusWhenAside = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.TREE_KEEP_FOCUS_WHEN_ASIDE, false);
             // * Server and connection automation
@@ -132,8 +150,10 @@ export class Config implements ConfigMembers {
             this.connectionPort = vscode.workspace.getConfiguration(Constants.CONFIG_NAME).get(Constants.CONFIG.IP_PORT, Constants.TCPIP_DEFAULT_PORT); // 32125
             // * Set context for tree items visibility that are based on config options
             this._leoIntegration.sendConfigToServer(this.getConfig());
+            utils.setContext(Constants.CONTEXT_FLAGS.LEO_TREE_BROWSE, this.leoTreeBrowse);
             utils.setContext(Constants.CONTEXT_FLAGS.TREE_IN_EXPLORER, this.treeInExplorer);
             utils.setContext(Constants.CONTEXT_FLAGS.SHOW_OPEN_ASIDE, this.showOpenAside);
+            utils.setContext(Constants.CONTEXT_FLAGS.SHOW_EDIT, this.showEditOnNodes);
             utils.setContext(Constants.CONTEXT_FLAGS.SHOW_ARROWS, this.showArrowsOnNodes);
             utils.setContext(Constants.CONTEXT_FLAGS.SHOW_ADD, this.showAddOnNodes);
             utils.setContext(Constants.CONTEXT_FLAGS.SHOW_MARK, this.showMarkOnNodes);