+++ src/leoIntegration.ts
@@ -14,6 +14,7 @@ import { LeoStatusBar } from "./leoStatusBar";
 import { CommandStack } from "./commandStack";
 import { LeoDocumentsProvider } from "./leoDocuments";
 import { LeoDocumentNode } from "./leoDocumentNode";
+import { LeoStates } from "./leoStates";
 
 /**
  * * Orchestrates Leo integration into vscode with treeview and file system providers
@@ -25,72 +26,8 @@ export class LeoIntegration {
     private _leoBridgeReadyPromise: Promise<LeoBridgePackage> | undefined; // Set when leoBridge has a leo controller ready
     private _currentOutlineTitle: string = Constants.GUI.TREEVIEW_TITLE_INTEGRATION; // Title has to be kept because it might need to be set (again) when either tree is first shown when switching visibility
 
-    private _leoBridgeReady: boolean = false; // Used along with executeCommand 'setContext' with Constants.CONTEXT_FLAGS.BRIDGE_READY
-    get leoBridgeReady(): boolean {
-        return this._leoBridgeReady;
-    }
-    set leoBridgeReady(p_value: boolean) {
-        this._leoBridgeReady = p_value;
-        utils.setContext(Constants.CONTEXT_FLAGS.BRIDGE_READY, p_value);
-    }
-
-    private _fileOpenedReady: boolean = false; // Used along with executeCommand 'setContext' with Constants.CONTEXT_FLAGS.TREE_OPENED
-    get fileOpenedReady(): boolean {
-        return this._fileOpenedReady;
-    }
-    set fileOpenedReady(p_value: boolean) {
-        this._fileOpenedReady = p_value;
-        utils.setContext(Constants.CONTEXT_FLAGS.TREE_OPENED, p_value);
-        this._setTreeViewTitle(
-            p_value ? Constants.GUI.TREEVIEW_TITLE : Constants.GUI.TREEVIEW_TITLE_INTEGRATION
-        );
-    }
-
-    // * States used in UI button visibility: changed, can undo/redo, can demote and can dehoist.
-    private _leoChanged: boolean = false;
-    get leoChanged(): boolean {
-        return this._leoChanged;
-    }
-    set leoChanged(p_value: boolean) {
-        this._leoChanged = p_value;
-        utils.setContext(Constants.CONTEXT_FLAGS.LEO_CHANGED, p_value);
-    }
-
-    private _leoCanUndo: boolean = false;
-    get leoCanUndo(): boolean {
-        return this._leoCanUndo;
-    }
-    set leoCanUndo(p_value: boolean) {
-        this._leoCanUndo = p_value;
-        utils.setContext(Constants.CONTEXT_FLAGS.LEO_CAN_UNDO, p_value);
-    }
-
-    private _leoCanRedo: boolean = false;
-    get leoCanRedo(): boolean {
-        return this._leoCanRedo;
-    }
-    set leoCanRedo(p_value: boolean) {
-        this._leoCanRedo = p_value;
-        utils.setContext(Constants.CONTEXT_FLAGS.LEO_CAN_REDO, p_value);
-    }
-
-    private _leoCanDemote: boolean = false;
-    get leoCanDemote(): boolean {
-        return this._leoCanDemote;
-    }
-    set leoCanDemote(p_value: boolean) {
-        this._leoCanDemote = p_value;
-        utils.setContext(Constants.CONTEXT_FLAGS.LEO_CAN_DEMOTE, p_value);
-    }
-
-    private _leoCanDehoist: boolean = false;
-    get leoCanDehoist(): boolean {
-        return this._leoCanDehoist;
-    }
-    set leoCanDehoist(p_value: boolean) {
-        this._leoCanDehoist = p_value;
-        utils.setContext(Constants.CONTEXT_FLAGS.LEO_CAN_DEHOIST, p_value);
-    }
+    // * State flags
+    public leoStates: LeoStates;
 
     // * Frontend command stack
     private _commandStack: CommandStack;
@@ -145,6 +82,7 @@ export class LeoIntegration {
     // * Outline Pane redraw/refresh flags. Also set when calling refreshTreeRoot
     // If there's no reveal and its the selected node, the old id will be re-used for the node. (see _id property in LeoNode)
     private _revealType: RevealType = RevealType.NoReveal; // to be read/cleared in arrayToLeoNodesArray, to check if any should self-select
+    private _preventShowBody = false; // Used when refreshing treeview from config: It requires not to open the body pane when refreshing.
 
     // * Documents Pane
     private _leoDocumentsProvider: LeoDocumentsProvider;
@@ -196,6 +134,9 @@ export class LeoIntegration {
     };
 
     constructor(private _context: vscode.ExtensionContext) {
+        // * Setup States
+        this.leoStates = new LeoStates(_context, this);
+
         // * Get configuration settings
         this.config = new Config(_context, this);
         this.config.buildFromSavedSettings();
@@ -320,7 +261,7 @@ export class LeoIntegration {
      * * Initiate a connection to the leoBridge server, then show view title, log pane, and set 'bridge ready' flags.
      */
     public connect(): void {
-        if (this.leoBridgeReady || this._leoIsConnecting) {
+        if (this.leoStates.leoBridgeReady || this._leoIsConnecting) {
             vscode.window.showInformationMessage(Constants.USER_MESSAGES.ALREADY_CONNECTED);
             return;
         }
@@ -332,7 +273,7 @@ export class LeoIntegration {
                 if (p_package.id !== 1) {
                     this.cancelConnect(Constants.USER_MESSAGES.CONNECT_ERROR);
                 } else {
-                    this.leoBridgeReady = true;
+                    this.leoStates.leoBridgeReady = true;
                     utils.setContext(Constants.CONTEXT_FLAGS.BRIDGE_READY, true);
                     this.showLogPane();
                     if (!this.config.connectToServerAutomatically) {
@@ -356,13 +297,13 @@ export class LeoIntegration {
      */
     public cancelConnect(p_message?: string): void {
         // 'disconnect error' versus 'failed to connect'
-        if (this.leoBridgeReady) {
+        if (this.leoStates.leoBridgeReady) {
             vscode.window.showErrorMessage(p_message ? p_message : Constants.USER_MESSAGES.DISCONNECTED);
         } else {
             vscode.window.showInformationMessage(p_message ? p_message : Constants.USER_MESSAGES.DISCONNECTED);
         }
-        this.fileOpenedReady = false;
-        this.leoBridgeReady = false;
+        this.leoStates.fileOpenedReady = false;
+        this.leoStates.leoBridgeReady = false;
         this._leoBridgeReadyPromise = undefined;
         this._leoStatusBar.update(false);
         this._refreshOutline(RevealType.RevealSelect);
@@ -388,7 +329,7 @@ export class LeoIntegration {
      * @param p_config A config object containing all the configuration settings
      */
     public sendConfigToServer(p_config: ConfigMembers): void {
-        if (this.fileOpenedReady) {
+        if (this.leoStates.fileOpenedReady) {
             this.sendAction(Constants.LEOBRIDGE.APPLY_CONFIG, JSON.stringify(p_config)).then(p_package => {
                 // Back from applying configuration to leobridgeserver.py
             });
@@ -401,14 +342,9 @@ export class LeoIntegration {
     private _triggerGetStates(): void {
         // Debounced timer has triggered so perform getStates action
         this._leoDocumentsProvider.refreshTreeRoot();
-
         this.sendAction(Constants.LEOBRIDGE.GET_STATES).then((p_package: LeoBridgePackage) => {
             if (p_package.states) {
-                this.leoChanged = p_package.states.changed;
-                this.leoCanUndo = p_package.states.canUndo;
-                this.leoCanRedo = p_package.states.canRedo;
-                this.leoCanDemote = p_package.states.canDemote;
-                this.leoCanDehoist = p_package.states.canDehoist;
+                this.leoStates.leoStateFlags(p_package.states);
             }
         });
     }
@@ -437,7 +373,7 @@ export class LeoIntegration {
      * * Setup leoInteg's UI for having no opened Leo documents
      */
     private _setupNoOpenedLeoDocument(): void {
-        this.fileOpenedReady = false;
+        this.leoStates.fileOpenedReady = false;
         this._bodyTextDocument = undefined;
         this.lastSelectedNode = undefined;
         this._refreshOutline(RevealType.RevealSelectFocus);
@@ -466,10 +402,10 @@ export class LeoIntegration {
             this._bodyFileSystemStarted = true;
         }
         // * Startup flag
-        this.fileOpenedReady = true;
+        this.leoStates.fileOpenedReady = true;
         // * First valid redraw of tree along with the selected node and its body
         this._refreshOutline(RevealType.RevealSelectFocus); // p_revealSelection flag set
-        // this._setTreeViewTitle(Constants.GUI.TREEVIEW_TITLE); // ? Maybe unused when used with welcome content
+        // this.setTreeViewTitle(Constants.GUI.TREEVIEW_TITLE); // ? Maybe unused when used with welcome content
         // * First StatusBar appearance
         this._leoStatusBar.show(); // Just selected a node
         // * Show leo log pane
@@ -518,9 +454,9 @@ export class LeoIntegration {
      * @param p_explorerView Flag to signify that the treeview who triggered this event is the one in the explorer view
      */
     private _onTreeViewVisibilityChanged(p_event: vscode.TreeViewVisibilityChangeEvent, p_explorerView: boolean): void {
-        if (p_event.visible && this.lastSelectedNode) {
+        if (p_event.visible) {
             this._lastVisibleTreeView = p_explorerView ? this._leoTreeExplorerView : this._leoTreeStandaloneView;
-            this._setTreeViewTitle();
+            this.setTreeViewTitle();
             this._needLastSelectedRefresh = true; // Its a new node in a new tree so refresh lastSelectedNode too
             this._refreshOutline(RevealType.RevealSelect);
         }
@@ -666,6 +602,32 @@ export class LeoIntegration {
     }
 
     /**
+     * * Show the outline, with Leo's selected node also selected, and optionally focussed
+     * @param p_focusOutline Flag for focus to be placed in outline
+     */
+    public showOutline(p_focusOutline: boolean): void {
+        if (this._lastSelectedNode) {
+            this._lastVisibleTreeView.reveal(this._lastSelectedNode, {
+                select: true,
+                focus: p_focusOutline
+            }).then(() => {
+                // console.log('done showing/revealing outline');
+            });
+        }
+    }
+
+    /**
+     * * Refresh tree for node hover icons refresh only
+     */
+    public configTreeRefresh(): void {
+        if (this.leoStates.fileOpenedReady && this.lastSelectedNode) {
+            this._revealType = RevealType.RevealSelect;
+            this._preventShowBody = true;
+            this._leoTreeDataProvider.refreshTreeRoot();
+        }
+    }
+
+    /**
      * * Refreshes the outline. A reveal type can be passed along to specify the reveal type for the selected node
      * @param p_revealType Facultative reveal type to specify type of reveal when the 'selected node' is encountered
      */
@@ -673,6 +635,7 @@ export class LeoIntegration {
         if (p_revealType !== undefined) { // To check if selected node should self-select while redrawing whole tree
             this._revealType = p_revealType; // To be read/cleared (in arrayToLeoNodesArray instead of directly by nodes)
         }
+        // Force showing last used Leo outline first
         if (this._lastSelectedNode && !(this._leoTreeExplorerView.visible || this._leoTreeStandaloneView.visible)) {
             this._lastVisibleTreeView.reveal(this._lastSelectedNode).then(() => {
                 this._leoTreeDataProvider.refreshTreeRoot();
@@ -721,6 +684,7 @@ export class LeoIntegration {
      * @param p_leoNode The node that was detected as the selected node in Leo
      */
     private _apToLeoNodeConvertReveal(p_leoNode: LeoNode): void {
+        this.leoStates.selectedNodeFlags(p_leoNode);
         // First setup flags for selecting and focusing based on the current reveal type needed
         const w_selectFlag = this._revealType >= RevealType.RevealSelect; // at least RevealSelect
         let w_focusFlag = this._revealType >= RevealType.RevealSelectFocus;  // at least RevealSelectFocus
@@ -833,6 +797,7 @@ export class LeoIntegration {
         if (p_aside && p_node !== this.lastSelectedNode) {
             this._revealTreeViewNode(p_node, { select: true, focus: false }); // no need to set focus: tree selection is set to right-click position
         }
+        this.leoStates.selectedNodeFlags(p_node);
         // TODO : #39 @boltex Save and restore selection, along with cursor position, from selection state saved in each node (or gnx array)
         this._leoStatusBar.update(true); // Just selected a node directly, or via expand/collapse
         const w_showBodyKeepFocus = p_aside ? this.config.treeKeepFocusWhenAside : this.config.treeKeepFocus;
@@ -887,6 +852,10 @@ export class LeoIntegration {
      * @param p_force_open Forces opening the body pane editor
      */
     private _showBodyIfRequired(p_aside: boolean, p_showBodyKeepFocus: boolean, p_force_open?: boolean): Thenable<vscode.TextEditor> {
+        if (this._preventShowBody) {
+            this._preventShowBody = false;
+            return Promise.resolve(vscode.window.activeTextEditor!);
+        }
         if (true || p_force_open || this._leoTreeStandaloneView.visible) {
             return this.showBody(p_aside, p_showBodyKeepFocus); // ! Always true for now to stabilize refreshes after derived files refreshes and others.
         } else {
@@ -1119,9 +1088,9 @@ export class LeoIntegration {
      * @param p_fromOutlineSignifies that the focus was, and should be brought back to, the outline
      */
     public saveAsLeoFile(p_fromOutline?: boolean): void {
-        if (!this.fileOpenedReady || this._isBusy()) { return; } // Warn user to wait for end of busy state
+        if (!this.leoStates.fileOpenedReady || this._isBusy()) { return; } // Warn user to wait for end of busy state
         // TODO : Implement & support multiple simultaneous files
-        if (this.fileOpenedReady) {
+        if (this.leoStates.fileOpenedReady) {
             if (this.lastSelectedNode) {
                 this.triggerBodySave(true)
                     .then(() => {
@@ -1143,9 +1112,9 @@ export class LeoIntegration {
      * @param p_fromOutlineSignifies that the focus was, and should be brought back to, the outline
      */
     public saveLeoFile(p_fromOutline?: boolean): void {
-        if (!this.fileOpenedReady || this._isBusy()) { return; } // Warn user to wait for end of busy state
+        if (!this.leoStates.fileOpenedReady || this._isBusy()) { return; } // Warn user to wait for end of busy state
         // TODO : Specify which file when supporting multiple simultaneous opened Leo files
-        if (this.fileOpenedReady) {
+        if (this.leoStates.fileOpenedReady) {
             if (this.lastSelectedNode && this._isCurrentFileNamed()) {
                 this.triggerBodySave(true)
                     .then(() => {
@@ -1163,7 +1132,7 @@ export class LeoIntegration {
      * * Show switch document 'QuickPick' dialog and switch file if selection is made, or just return if no files are opened.
      */
     public switchLeoFile(): void {
-        if (!this.fileOpenedReady || this._isBusy()) { return; } // Warn user to wait for end of busy state
+        if (!this.leoStates.fileOpenedReady || this._isBusy()) { return; } // Warn user to wait for end of busy state
         // get list and show dialog to user, even if there's only one...but at least one!
         // TODO : p_package members names should be made into constants
         // TODO : Fix / Cleanup opened body panes close before reopening when switching
@@ -1223,7 +1192,7 @@ export class LeoIntegration {
      */
     public closeLeoFile(): void {
         if (this._isBusy()) { return; } // Warn user to wait for end of busy state
-        if (this.fileOpenedReady) {
+        if (this.leoStates.fileOpenedReady) {
             this.triggerBodySave(true)
                 .then(() => {
                     return this.sendAction(Constants.LEOBRIDGE.CLOSE_FILE, JSON.stringify({ forced: false }));
@@ -1337,7 +1306,7 @@ export class LeoIntegration {
      * * Set the outline pane top bar string message
      * @param p_title new string to replace the current title
      */
-    private _setTreeViewTitle(p_title?: string): void {
+    public setTreeViewTitle(p_title?: string): void {
         if (p_title) {
             this._currentOutlineTitle = p_title;
         }
@@ -1361,7 +1330,7 @@ export class LeoIntegration {
      * * StatusBar click handler
      */
     public statusBarOnClick(): void {
-        if (this.fileOpenedReady) {
+        if (this.leoStates.fileOpenedReady) {
             this.switchLeoFile();
         } else {
             this.showLeoCommands();